+++ src/main/java/com/project/domain/webpush/dto/request/AdminPushRequest.java
@@ -1,4 +1,4 @@
-package com.project.domain.webpush.controller.dto;
+package com.project.domain.webpush.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
