+++ src/main/java/com/project/domain/webpush/service/WebPushServiceImpl.java
@@ -21,7 +21,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.domain.family.repository.FamilyMemberRepository;
-import com.project.domain.webpush.controller.dto.PushSubscriptionRequest;
+import com.project.domain.webpush.dto.request.PushSubscriptionRequest;
 import com.project.domain.webpush.entity.PushSubscription;
 import com.project.domain.webpush.repository.PushSubscriptionRepository;
 import com.project.global.exception.ApplicationException;
